# execution: reuse iterator in hash join