--- a/util/chunk/iterator.go
+++ b/util/chunk/iterator.go
@@ -17,7 +17,7 @@ var (
 	_ Iterator = (*Iterator4Chunk)(nil)
 	_ Iterator = (*iterator4RowPtr)(nil)
 	_ Iterator = (*iterator4List)(nil)
-	_ Iterator = (*iterator4Slice)(nil)
+	_ Iterator = (*Iterator4Slice)(nil)
 	_ Iterator = (*iterator4RowContainer)(nil)
 	_ Iterator = (*multiIterator)(nil)
 )
@@ -52,16 +52,23 @@ type Iterator interface {
 
 // NewIterator4Slice returns a Iterator for Row slice.
 func NewIterator4Slice(rows []Row) Iterator {
-	return &iterator4Slice{rows: rows}
+	return &Iterator4Slice{rows: rows}
 }
 
-type iterator4Slice struct {
+// Iterator4Slice is used to iterate rows inside a slice.
+type Iterator4Slice struct {
 	rows   []Row
 	cursor int
 }
 
+// Reset reset by rows
+func (it *Iterator4Slice) Reset(rows []Row) {
+	it.rows = rows
+	it.cursor = 0
+}
+
 // Begin implements the Iterator interface.
-func (it *iterator4Slice) Begin() Row {
+func (it *Iterator4Slice) Begin() Row {
 	if it.Len() == 0 {
 		return it.End()
 	}
@@ -70,7 +77,7 @@ func (it *iterator4Slice) Begin() Row {
 }
 
 // Next implements the Iterator interface.
-func (it *iterator4Slice) Next() Row {
+func (it *Iterator4Slice) Next() Row {
 	if len := it.Len(); it.cursor >= len {
 		it.cursor = len + 1
 		return it.End()
@@ -81,7 +88,7 @@ func (it *iterator4Slice) Next() Row {
 }
 
 // Current implements the Iterator interface.
-func (it *iterator4Slice) Current() Row {
+func (it *Iterator4Slice) Current() Row {
 	if it.cursor == 0 || it.cursor > it.Len() {
 		return it.End()
 	}
@@ -89,22 +96,22 @@ func (it *iterator4Slice) Current() Row {
 }
 
 // End implements the Iterator interface.
-func (it *iterator4Slice) End() Row {
+func (it *Iterator4Slice) End() Row {
 	return Row{}
 }
 
 // ReachEnd implements the Iterator interface.
-func (it *iterator4Slice) ReachEnd() {
+func (it *Iterator4Slice) ReachEnd() {
 	it.cursor = it.Len() + 1
 }
 
 // Len implements the Iterator interface.
-func (it *iterator4Slice) Len() int {
+func (it *Iterator4Slice) Len() int {
 	return len(it.rows)
 }
 
 // Error returns none-nil error if anything wrong happens during the iteration.
-func (it *iterator4Slice) Error() error {
+func (it *Iterator4Slice) Error() error {
 	return nil
 }