### What problem does this PR solve?

Issue Number: https://github.com/pingcap/tidb/issues/14466

Problem Summary:

the method joinMatchedProbeSideRow2Chunk uses the method NewIterator4Slice to create a lot of iterators whereas these iterators can be reused.

### What is changed and how it works?

How it Works:

reset iterator with ``[]Row`` when using it

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Integration test

```
name                                                                                                        old time/op  new time/op  delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-6   1.18s ± 2%   1.18s ± 3%  +0.32%  (p=0.739 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-6     249ms ± 1%   247ms ± 0%  -0.52%  (p=0.004 n=9+9)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-6      9.15s ± 2%   9.10s ± 3%  -0.53%  (p=0.278 n=9+10)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-6                72.7µs ± 2%  72.0µs ± 3%  -0.86%  (p=0.247 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-6         32.6ms ± 3%  32.5ms ± 3%  -0.51%  (p=0.247 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-6           29.9ms ± 2%  30.2ms ± 1%  +0.99%  (p=0.089 n=10+10)
```

Side effects

- Performance regression
    - Consumes less CPU
    - Consumes lessMEM

### Release note <!-- bugfixes or new feature need a release note -->

- execution: reuse iterator in hash join