@tisonkun 2021-07-14
/cc @tiancaiamao 

I think it may help on performance which you focus on now.