@tiancaiamao 2021-07-16
![image](https://user-images.githubusercontent.com/1420062/125882475-5a70c472-249f-4a58-887a-86101a23c51e.png)

It's a relatively small part if you see the whole picture. @jianzhiyao 
I'm not saying this change is meaningless, but I meant to focus on something more important.

The more close to the user level, the better improvement an optimization can achieve.
Understanding the code, understanding the scenario.

For example, with correct statistics and choose the correct plan, the speed can be 50x 100x difference.
A better architecture, or a better algorithm for a specific problem can be 2-5x or even 10x difference.
The low level optimization, however, take a lot of work, and maybe improve at most 30%-50%

Plus, I'm not familiar enough with this part of code, so   @wjhuang2016 @ichn-hu or maybe @qw4990  PTAL


> /cc @tiancaiamao
> 
> I think it may help on performance which you focus on now.

Just some work on point get ...  @tisonkun 
Because that's 80% of the OLTP case, and it's simple enough, hard to optimze from the other directions except for the low level.

What I really care about is our benchmark system... we need the metrics and keep it running  #25754
(as I can see it's not built yet, although there are some benchmark tests do exist)
Before we can really improve something, we need to have better understanding what the whole picture is now.